requirements.txt
Django==5.1.3
djangorestframework==3.15.2
mysqlclient==2.2.5
django-mysql==4.15.0
django-cors-headers==4.6.0
pika==1.3.2
requests==2.32.3
